+++ tests/StacktraceTest.php
@@ -203,7 +203,7 @@ final class StacktraceTest extends TestCase
             ],
         ];
 
-        yield 'Backtrace with frame containing memory address' => [
+        yield 'Backtrace with frame containing memory address in PHP <7.4.2 format' => [
             new Options([
                 'prefixes' => ['/path-prefix'],
             ]),
@@ -218,6 +218,45 @@ final class StacktraceTest extends TestCase
                     ],
                     [
                         'class' => "class@anonymous\x00/path-prefix/path/to/app/consumer.php0x7fc3bc369418",
+                        'function' => 'messageCallback',
+                        'type' => '->',
+                    ],
+                ],
+            ],
+            [
+                [
+                    null,
+                    Frame::INTERNAL_FRAME_FILENAME,
+                    0,
+                ],
+                [
+                    "class@anonymous\x00/path/to/app/consumer.php::messageCallback",
+                    Frame::INTERNAL_FRAME_FILENAME,
+                    0,
+                ],
+                [
+                    "class@anonymous\x00/path/to/app/consumer.php::messageCallback",
+                    'path/to/file',
+                    12,
+                ],
+            ],
+        ];
+
+        yield 'Backtrace with frame containing memory address in PHP >=7.4.2 format' => [
+            new Options([
+                'prefixes' => ['/path-prefix'],
+            ]),
+            [
+                'file' => 'path/to/file',
+                'line' => 12,
+                'backtrace' => [
+                    [
+                        'class' => "class@anonymous\x00/path/to/app/consumer.php:12$3e0a7",
+                        'function' => 'messageCallback',
+                        'type' => '->',
+                    ],
+                    [
+                        'class' => "class@anonymous\x00/path-prefix/path/to/app/consumer.php:12$3e0a7",
                         'function' => 'messageCallback',
                         'type' => '->',
                     ],